+++ APIBlox.AspNetCore.DynamicControllers/Extensions/ServiceCollectionExtensions.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Reflection;
 using APIBlox.AspNetCore;
 using APIBlox.AspNetCore.Contracts;
-using APIBlox.AspNetCore.Exceptions;
 using APIBlox.NetCore.Extensions;
 using Microsoft.Extensions.Logging;
 
@@ -18,18 +17,22 @@ namespace Microsoft.Extensions.DependencyInjection
     public static class ServiceCollectionExtensions
     {
         /// <summary>
-        ///     Adds and configures dynamic controller configurations using the <see cref="DynamicControllerFactory"/>.
+        ///     Adds and configures dynamic controller configurations using the <see cref="DynamicControllerFactory" />.
         /// </summary>
         /// <param name="services">The services.</param>
         /// <param name="loggerFactory">The logger factory.</param>
         /// <param name="startup">The startup type.</param>
         /// <param name="useCached">if set to <c>true</c> [use caching].</param>
+        /// <param name="addControllerComments">
+        ///     if set to <c>true</c> Summary comments from the request object of the first
+        ///     controller action added will be included as controller summary comments..
+        /// </param>
         /// <param name="preCompile">The pre compile function.</param>
         /// <param name="postCompile">The post compile action.</param>
         /// <param name="xmlFallbackPaths">The XML fallback paths.</param>
         /// <returns>IServiceCollection.</returns>
         public static IServiceCollection AddDynamicControllerConfigurations(this IServiceCollection services,
-            ILoggerFactory loggerFactory, Type startup, bool useCached,
+            ILoggerFactory loggerFactory, Type startup, bool useCached, bool addControllerComments,
             Func<DynamicControllerFactory, IEnumerable<IComposedTemplate>> preCompile,
             Action<DynamicControllerFactory, string, Assembly> postCompile,
             params string[] xmlFallbackPaths
@@ -43,9 +46,10 @@ namespace Microsoft.Extensions.DependencyInjection
 
             var factory = new DynamicControllerFactory(loggerFactory,
                 assemblyName,
-                useCached
+                useCached,
+                addControllerComments
             );
-            
+
             var templates = preCompile?.Invoke(factory);
 
             if (templates is null)